Some Changes

+++ Domain.Contracts/DTOs/DrinkItemDto.cs
@@ -1,8 +1,7 @@
-﻿
-public class DrinkItemDto
+﻿public class DrinkItemDto
 {
     public int Id { get; set; }
     public string Name { get; set; }
-    public int Price { get; set; }
+    public decimal Price { get; set; }
 }
